web/requirements.txt
asgiref==3.6.0
crispy-bootstrap5==2024.2
Django==4.2.10
django-crispy-forms==2.1
djangorestframework==3.15.1
flake8==6.0.0
gunicorn==21.0.1
mccabe==0.7.0
mssql-django==1.3
packaging==23.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pyodbc==4.0.39
pytz==2023.3
sqlparse==0.4.4
